iwinfo: dot11ah: implement iwinfo backend to unroll Morse Micro dot11ah shim
Morse Micro devices currently use a shim driver to report 802.11ah channel
information using existing 802.11ac channels. This commit adds a set of
dot11ah ops into the nl80211 iwinfo backend to unravel this shim and correctly
report HaLow frequencies for Morse Micro devices.

A HaLow AP may send short beacons. In order to report the encryption type during
a passive scan, we infer it from the values in the RSNXE information element. As
HaLow only supports WPA3, we do this by checking the hash-to-element field.
This is not a robust, long term solution.

Simple RSNXE parsing added as required.

Co-authored-by: Milad Mohtashamirad <milad.mohtashamirad@morsemicro.com>
Co-authored-by: Arien Judge <arien.judge@morsemicro.com>