nl80211: (APP-5631, PR #37) fix corrupt scan results from wpa_supplicant
Without clearing the scanlist entries, it was possible to have bad
data remaining from previous calls in the ht_operation/vht_operation
fields (not set here).  This could cause segfaults in rpcd's iwinfo
module, as it was possible that (a) the primary_chan would not be 0
and therefore (b) you would use bad data for chan_width etc.  as an
index into short-ish array:

	if (e->ht_chan_info.primary_chan) {
		t = blobmsg_open_table(&buf, "ht_operation");
		blobmsg_add_u32(&buf, "primary_channel", e->ht_chan_info.primary_chan);
		blobmsg_add_string(&buf, "secondary_channel_offset", ht_secondary_offset[e->ht_chan_info.secondary_chan_off]);
		blobmsg_add_u32(&buf, "channel_width", ht_chan_width[e->ht_chan_info.chan_width]);
		blobmsg_close_table(&buf, t);
	}

See also:

https://github.com/openwrt/iwinfo/pull/30


Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge