nl80211: (APP-5768, PR #38) support MM6108A2 and any new chip
We should gate our probe on whether the module is being used
(this will also help with native s1g), which would mean this
chip would come up ok regardless, but also add its chip id.

Also, since we know it's a Morse chip, we may as well say this
if the chip id isn't there.

We use the module rather than the driver so we minimise
the chance of a native s1g accidental detection, even though
this is 'bad' (what if it's not built as a module?).


Approved-by: Sophronia Koilpillai
Approved-by: Andrew Dodd
Approved-by: Lyall Beveridge