nl80211: (APP-5838) skip EU for standard nl80211
EU was added to the main list so that other parts of the system
(e.g. rpcd, iwinfo cli) could look up the country.
Unfortunately, this means nl80211_get_countrylist can no
longer cheerfully report _all_ the countries (IMO, it really
should be looking at what's currently in wireless-regdb, but
this is difficult to extract outside the kernel and I don't
believe the kernel has a command to list the currently available
countries).