nl80211: (APP-3770, PR #39) return better encryption info for HaLow APs
S1G short beacons do not have sufficient information to determine the
encryption, so we default to active scans to increase the
chance that we get good info.

This also implements the wpa_supplicant handling so that
if we have an existing connection we ask wpa_supplicant to do
the scan instead, AND if we don't get any good info we simply
claim that it supports everything.

We do not do an active scan while hostapd_s1g is up to avoid
disruption. Unfortunately, this involves a patch to the normal
iwinfo_nl80211 code.


Approved-by: Sophronia Koilpillai
Approved-by: Prabha Murugesan