dot11ah: (APP-5955, PR #41) fix channel/bandwidth reporting
- stop using wpa_supplicant
  (does not return sufficient information to determine channel)
- add the missing JP/AU channels (overlapping channels)
- use VHT/HT info to determine bandwidth and make logic there
  simpler

* dot11ah: (APP-5955) simplify logic

Don't get into the ht channel info at all, and always simply
use the main channel for the primary channel to simplifiy
logic (which should be the 1MHz channel).

Be explicit about what happens when we don't the VHT Oper IE
(i.e. report 0 for the channel, which will lead to unknown).


Approved-by: Sophronia Koilpillai
Approved-by: Prabha Murugesan
Approved-by: Lyall Beveridge